pyproject.toml
[project]
name = "linkedin_mcp"
version = "0.0.1"
description = "A LinkedIn MCP based on Restack"
authors = [{ name = "Restack Team", email = "service@restack.io" }]
requires-python = ">=3.10,<3.14"
readme = "README.md"
dependencies = [
    "pydantic>=2.10.6",
    "watchfiles>=1.0.4",
    "python-dotenv==1.0.1",
    "openai>=1.61.0",
    "brightdata-sdk>=1.1.3",
    "restack-ai>=0.0.114",
]

[project.scripts]
dev = "src.services:watch_services"
services = "src.services:run_services"
schedule = "schedule:run_schedule"

[tool.hatch.build.targets.sdist]
include = ["src"]

[tool.hatch.build.targets.wheel]
include = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
